requirements.txt
numpy==1.25.2
browserhistory==0.1.2
opencv_python==4.9.0.80
Pillow==10.4.0
pycryptodome==3.20.0
pynput==1.7.6
pyperclip==1.8.2
Requests==2.32.3
scipy==1.13.0
sounddevice==0.4.6
